Refactor first round implementation and split into files
